refractor code into different files
